+++ src/main/java/com/solace/spring/cloud/stream/binder/config/SolaceHealthIndicatorsConfiguration.java
@@ -5,14 +5,12 @@ import com.solace.spring.cloud.stream.binder.health.contributors.BindingsHealthC
 import com.solace.spring.cloud.stream.binder.health.contributors.SolaceBinderHealthContributor;
 import com.solace.spring.cloud.stream.binder.health.handlers.SolaceSessionEventHandler;
 import com.solace.spring.cloud.stream.binder.health.indicators.SessionHealthIndicator;
-import com.solace.spring.cloud.stream.binder.properties.SolaceSessionHealthProperties;
 import com.solacesystems.jcsmp.JCSMPProperties;
 import com.solacesystems.jcsmp.SolaceSessionOAuth2TokenProvider;
 import jakarta.annotation.Nullable;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.actuate.autoconfigure.health.ConditionalOnEnabledHealthIndicator;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -20,7 +18,6 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @ConditionalOnClass(name = "org.springframework.boot.actuate.health.HealthIndicator")
 @ConditionalOnEnabledHealthIndicator("binders")
-@EnableConfigurationProperties({SolaceSessionHealthProperties.class})
 public class SolaceHealthIndicatorsConfiguration {
 
     @Bean
@@ -30,13 +27,12 @@ public class SolaceHealthIndicatorsConfiguration {
     }
 
     @Bean
-    public SolaceBinderHealthContributor solaceBinderHealthContributor(
-            SolaceSessionHealthProperties solaceSessionHealthProperties) {
+    public SolaceBinderHealthContributor solaceBinderHealthContributor() {
         if (log.isDebugEnabled()) {
             log.debug("Creating Solace Connection Health Indicators Hierarchy");
         }
         return new SolaceBinderHealthContributor(
-                new SessionHealthIndicator(solaceSessionHealthProperties),
+                new SessionHealthIndicator(),
                 new BindingsHealthContributor()
         );
     }