+++ src/test/java/com/solace/spring/cloud/stream/binder/health/indicators/SolaceBinderHealthIndicatorTest.java
@@ -1,43 +1,42 @@
 package com.solace.spring.cloud.stream.binder.health.indicators;
 
-import com.solace.spring.cloud.stream.binder.properties.SolaceSessionHealthProperties;
-import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.actuate.health.Health;
 import org.springframework.boot.actuate.health.Status;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class SolaceBinderHealthIndicatorTest {
     private SessionHealthIndicator healthIndicator;
 
     @BeforeEach
     void BeforeEach() {
-        SolaceSessionHealthProperties solaceSessionHealthProperties = new SolaceSessionHealthProperties();
-        solaceSessionHealthProperties.setReconnectAttemptsUntilDown(10);
-        this.healthIndicator = new SessionHealthIndicator(solaceSessionHealthProperties);
+        this.healthIndicator = new SessionHealthIndicator();
     }
 
     @Test
-    public void testUp(SoftAssertions softly) {
+    public void testUp() {
         healthIndicator.up();
         Health health = healthIndicator.health();
-        softly.assertThat(health.getStatus()).isEqualTo(Status.UP);
-        softly.assertThat(health.getDetails()).isEmpty();
+        assertEquals(Status.UP, health.getStatus());
+        assertTrue(health.getDetails().isEmpty());
     }
 
     @Test
-    public void testDown(SoftAssertions softly) {
+    public void testDown() {
         healthIndicator.down(null);
         Health health = healthIndicator.health();
-        softly.assertThat(health.getStatus()).isEqualTo(Status.DOWN);
-        softly.assertThat(health.getDetails()).isEmpty();
+        assertEquals(Status.DOWN, health.getStatus());
+        assertTrue(health.getDetails().isEmpty());
     }
 
     @Test
-    public void testReconnecting(SoftAssertions softly) {
+    public void testReconnecting() {
         healthIndicator.reconnecting(null);
         Health health = healthIndicator.health();
-        softly.assertThat(health.getStatus()).isEqualTo(new Status("RECONNECTING"));
-        softly.assertThat(health.getDetails()).isEmpty();
+        assertEquals(Status.DOWN, health.getStatus()); // reconnecting now goes to DOWN
+        assertTrue(health.getDetails().isEmpty());
     }
 }